+++ launch.py
@@ -1,6 +1,6 @@
 from argparse import ArgumentParser
-from colladamaya.maya import Maya
-from colladamax.max import Max
+from maya.maya import Maya
+from max.max import Max
 from common.opencollada import OpenCOLLADA
 from common.test_runner import TestRunner
 from common.util import *
