+++ maya/scripts/maya_export_script.py
