+++ maya/tests/camera/Cube.export_options
