+++ maya/tests/camera/Cube.mb
