+++ maya/tests/camera/TestCamera.py
